infra/transaction_bots/requirements.txt
web3
requests
python-bitcoinlib